feat: inbox switch button
