+++ apps/renderer/src/modules/feed-column/list.desktop.tsx
@@ -2,7 +2,6 @@ import { useDraggable } from "@dnd-kit/core"
 import { ScrollArea } from "@follow/components/ui/scroll-area/index.js"
 import { cn, isKeyForMultiSelectPressed } from "@follow/utils/utils"
 import { forwardRef, memo, useEffect, useImperativeHandle, useMemo, useRef, useState } from "react"
-import { useTranslation } from "react-i18next"
 import Selecto from "react-selecto"
 import { useEventListener } from "usehooks-ts"
 
@@ -17,35 +16,18 @@ import {
 } from "./atom"
 import { DraggableContext } from "./context"
 import { useShouldFreeUpSpace } from "./hook"
-import {
-  EmptyFeedList,
-  ListHeader,
-  StarredItem,
-  useFeedsGroupedData,
-  useInboxesGroupedData,
-  useListsGroupedData,
-} from "./list.shared"
+import { EmptyFeedList, ListHeader, StarredItem, useFeedsGroupedData } from "./list.shared"
 import { SortableFeedList } from "./sort-by"
 
 const FeedListImpl = forwardRef<HTMLDivElement, { className?: string; view: number }>(
   ({ className, view }, ref) => {
     const feedsData = useFeedsGroupedData(view)
-    const listsData = useListsGroupedData(view)
-    const inboxesData = useInboxesGroupedData(view)
     const categoryOpenStateData = useCategoryOpenStateByView(view)
 
-    const hasData =
-      Object.keys(feedsData).length > 0 ||
-      Object.keys(listsData).length > 0 ||
-      Object.keys(inboxesData).length > 0
-
-    const { t } = useTranslation()
+    const hasData = Object.keys(feedsData).length > 0
 
     // Data prefetch
     useAuthQuery(Queries.lists.list())
-
-    const hasListData = Object.keys(listsData).length > 0
-    const hasInboxData = Object.keys(inboxesData).length > 0
 
     const scrollerRef = useRef<HTMLDivElement>(null)
     const selectoRef = useRef<Selecto>(null)
@@ -216,16 +198,6 @@ const FeedListImpl = forwardRef<HTMLDivElement, { className?: string; view: numb
           <StarredItem view={view} />
           <DraggableContext.Provider value={draggableContextValue}>
             <div className="space-y-px" id="feeds-area" ref={setNodeRef}>
-              {(hasListData || hasInboxData) && (
-                <div
-                  className={cn(
-                    "mb-1 flex h-6 w-full shrink-0 items-center rounded-md px-2.5 text-xs font-semibold text-theme-vibrancyFg transition-colors",
-                    Object.keys(feedsData).length === 0 ? "mt-0" : "mt-1",
-                  )}
-                >
-                  {t("words.feeds")}
-                </div>
-              )}
               {hasData ? (
                 <SortableFeedList
                   view={view}
